Add RawBind function
Add a function to bind an IRODSConnection to an existing socket, without
further handling of authentication. This can be used for an irods proxy
server to listen on a socket, parse the first message which is always the
IRODSMessageStartupPack, open a connection to a backend based on
information in this startup message, and forward the startup message and
further communication to this backend. The user and zone are sent in the
startup message - so different zones could be served at a single ip
address.

Signed-off-by: Peter Verraedt <peter@verraedt.be>