Fix empty SpecColl_PI in DataObjInp_PI
Make SpecialCollectionPointer a pointer to the
IRODSMessageSpecialCollection struct to reflect the behaviour of the
C/python clients. In usual data object creation or modification
requests, the SpecColl_PI is empty, but we were not sending a nilpointer
but an empty xml struct. This caused that the corresponding dynamic peps
would contain an empty objPath in its DATAOBJINP argument (probably
because the objPath of the empty SpecColl_PI was overwriting the value
in the main struct).

Signed-off-by: Peter Verraedt <peter@verraedt.be>